+++ types/select/index.d.ts
@@ -85,7 +85,7 @@ export interface AutoCompleteProps extends HTMLAttributesWeak, CommonProps {
     /**
      * 弹层显示或隐藏时触发的回调
      */
-    onVisibleChange?: (visible: boolean) => void;
+    onVisibleChange?: (visible: boolean, type: string) => void;
 
     /**
      * 弹层挂载的容器节点
@@ -161,6 +161,16 @@ export interface AutoCompleteProps extends HTMLAttributesWeak, CommonProps {
      * 自动高亮第一个选项
      */
     autoHighlightFirstItem?: boolean;
+
+    /**
+     * 高亮key
+     */
+    highlightKey?: string;
+
+    /**
+     *  默认高亮key
+     */
+    defaultHighlightKey?: string;
 }
 
 export class AutoComplete extends React.Component<AutoCompleteProps, any> {}
@@ -440,6 +450,16 @@ export interface SelectProps extends Omit<HTMLAttributesWeak, 'renderPreview'>, 
      * 自动高亮第一个选项
      */
     autoHighlightFirstItem?: boolean;
+
+    /**
+     * 高亮key
+     */
+    highlightKey?: string;
+
+    /**
+     *  默认高亮key
+     */
+    defaultHighlightKey?: string;
 }
 
 export default class Select extends React.Component<SelectProps, any> {